+++ NEWS.md
@@ -4,6 +4,7 @@
 
 * Add `AIC.PlackettLuce` to enable computation of AIC on new observations (e.g.
 data held out in cross-validation).
+* Add `fitted.pltree` to return combined fitted probabilities for each choice within each ranking, for each node in a Plackett-Luce tree.
 
 ## Bug fixes